finish api authorizations

+++ src/app/api/deck-cards/[deckCardId]/route.ts
@@ -1,5 +1,6 @@
 import { NextResponse } from "next/server";
 import { prisma } from "../../../../../prisma/prisma";
+import { auth } from "@clerk/nextjs";
 
 export async function PATCH(
   request: Request,
@@ -14,7 +15,35 @@ export async function PATCH(
   try {
     const res = await request.json();
 
+    const { userId } = auth();
+
+    if (!userId) {
+      return NextResponse.json({ error: "Not authorized" }, { status: 401 });
+    }
+
     const deckCardId = params.deckCardId;
+
+    // Retrieve the existing deckCard with its associated deck
+    const existingDeckCard = await prisma.deckCard.findUnique({
+      where: {
+        id: deckCardId,
+      },
+      include: {
+        deck: true, // include the associated deck
+      },
+    });
+
+    // If the deckCard does not exist or the deck's userId does not match the authenticated userId
+    if (
+      !existingDeckCard ||
+      !existingDeckCard.deck ||
+      existingDeckCard.deck.userId !== userId
+    ) {
+      return NextResponse.json(
+        { error: "Not authorized to make this update" },
+        { status: 401 }
+      );
+    }
 
     if (res.quantity <= 0) {
       await prisma.deckCard.delete({
@@ -22,7 +51,7 @@ export async function PATCH(
           id: deckCardId,
         },
       });
-      return NextResponse.json({ message: "DeckCard deleted" });
+      return NextResponse.json({ message: "Deck card deleted" });
     } else {
       const updatedDeckCard = await prisma.deckCard.update({
         where: {
